More java updates.
