+++ java/src/main/java/edu/umn/nlpie/mtap/processing/EventProcessor.java
@@ -19,6 +19,7 @@ package edu.umn.nlpie.mtap.processing;
 import edu.umn.nlpie.mtap.common.JsonObject;
 import edu.umn.nlpie.mtap.common.JsonObjectBuilder;
 import edu.umn.nlpie.mtap.model.Event;
+import edu.umn.nlpie.mtap.model.ProtoLabelAdapter;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -44,97 +45,7 @@ import java.util.*;
  * ensuring thread-safety.
  */
 public abstract class EventProcessor extends ProcessorBase {
-  public static @NotNull Map<String, Object> metadataMap(Class<?> processorClass) {
-    Processor processor = processorClass.getAnnotation(Processor.class);
-    Map<String, Object> map = new LinkedHashMap<>();
-    map.put("name", processor.value());
-    if (!"".equals(processor.description())) {
-      map.put("description", processor.description());
-    }
-    String humanName = processor.humanName();
-    if ("".equals(humanName)) {
-      humanName = processorClass.getSimpleName();
-    }
-    map.put("human_name", humanName);
-
-    if (processor.inputs().length > 0) {
-      List<Map<String, Object>> inputs = new ArrayList<>();
-      for (LabelIndexDescription input : processor.inputs()) {
-        inputs.add(descToMap(input));
-      }
-      map.put("inputs", inputs);
-    }
-
-    if (processor.outputs().length > 0) {
-      List<Map<String, Object>> outputs = new ArrayList<>();
-      for (LabelIndexDescription output : processor.outputs()) {
-        outputs.add(descToMap(output));
-      }
-      map.put("outputs", outputs);
-    }
-
-    if (processor.parameters().length > 0) {
-      List<Map<String, Object>> parameters = new ArrayList<>();
-      for (ParameterDescription parameter : processor.parameters()) {
-        Map<String, Object> paramMap = new LinkedHashMap<>();
-        paramMap.put("name", parameter.name());
-        if (!"".equals(parameter.description())) {
-          paramMap.put("description", parameter.description());
-        }
-        paramMap.put("data_type", parameter.dataType());
-        paramMap.put("required", parameter.required());
-        parameters.add(paramMap);
-      }
-      map.put("parameters", parameters);
-    }
-
-    for (KeyValue keyValue : processor.additionalMetadata()) {
-      map.put(keyValue.key(), keyValue.value());
-    }
-
-    if (!map.containsKey("implementation_lang")) {
-      map.put("implementation_lang", "Java");
-    }
-
-    if (!map.containsKey("entry_point")) {
-      map.put("entry_point", processorClass.getCanonicalName());
-    }
-
-    return map;
-  }
-
-  private static @NotNull Map<String, Object> descToMap(LabelIndexDescription input) {
-    Map<String, Object> map = new LinkedHashMap<>();
-    map.put("name", input.name());
-    String reference = input.reference();
-    if (!"".equals(reference)) {
-      map.put("reference", reference);
-    }
-    String nameFromParameter = input.nameFromParameter();
-    if (!"".equals(nameFromParameter)) {
-      map.put("name_from_parameter", nameFromParameter);
-    }
-    if (!"".equals(input.description())) {
-      map.put("description", input.description());
-    }
-    if (input.properties().length > 0) {
-      List<Map<String, Object>> properties = new ArrayList<>();
-      for (PropertyDescription property : input.properties()) {
-        Map<String, Object> propertyMap = new LinkedHashMap<>();
-        propertyMap.put("name", property.name());
-        if (!"".equals(property.description())) {
-          propertyMap.put("description", property.description());
-        }
-        if (!"".equals(property.dataType())) {
-          propertyMap.put("data_type", property.dataType());
-        }
-        propertyMap.put("nullable", property.nullable());
-        properties.add(propertyMap);
-      }
-      map.put("properties", properties);
-    }
-    return map;
-  }
+  private final @NotNull Map<@NotNull String, @NotNull ProtoLabelAdapter<?>> defaultAdapters = new HashMap<>();
 
   /**
    * Returns a map of this processor's metadata as specified in the {@link Processor} annotation.
@@ -143,6 +54,33 @@ public abstract class EventProcessor extends ProcessorBase {
    */
   public @NotNull Map<@NotNull String, @Nullable Object> getProcessorMetadata() {
     return metadataMap(getClass());
+  }
+
+  /**
+   * Adds a default adapter for a specfied label index name.
+   * @param indexName The index name.
+   * @param adapter The adapter.
+   */
+  protected void addDefaultAdapter(@NotNull String indexName, @NotNull ProtoLabelAdapter<?> adapter) {
+    defaultAdapters.put(indexName, adapter);
+  }
+
+  /**
+   * Adds a map of default adapters.
+   *
+   * @param adapterMap A mapping from index names to label adapters.
+   */
+  protected void addAllDefaultAdapters(@NotNull Map<@NotNull String, @NotNull ProtoLabelAdapter<?>> adapterMap) {
+    this.defaultAdapters.putAll(adapterMap);
+  }
+
+  /**
+   * An unmodifiable view of the default label adapters.
+   *
+   * @return default adapters map.
+   */
+  public @NotNull Map<@NotNull String, @NotNull ProtoLabelAdapter<?>> getDefaultAdapters() {
+    return Collections.unmodifiableMap(defaultAdapters);
   }
 
   /**
@@ -155,7 +93,7 @@ public abstract class EventProcessor extends ProcessorBase {
   public abstract void process(
       @NotNull Event event,
       @NotNull JsonObject params,
-      @NotNull JsonObjectBuilder result
+      @NotNull JsonObjectBuilder<?, ?> result
   );
 
   /**
