+++ java/src/test/java/edu/umn/nlpie/mtap/processing/EventProcessorTest.java
@@ -17,7 +17,6 @@
 package edu.umn.nlpie.mtap.processing;
 
 import edu.umn.nlpie.mtap.model.Event;
-import edu.umn.nlpie.mtap.model.EventBuilder;
 import edu.umn.nlpie.mtap.common.JsonObject;
 import edu.umn.nlpie.mtap.common.JsonObjectBuilder;
 import edu.umn.nlpie.mtap.common.JsonObjectImpl;
@@ -34,7 +33,7 @@ public class EventProcessorTest {
   @Test
   void startedStopwatchNoContext() {
     WithStartedStopwatch withStartedStopwatch = new WithStartedStopwatch();
-    withStartedStopwatch.process(EventBuilder.newBuilder().build(), JsonObjectImpl.newBuilder().build(), JsonObjectImpl.newBuilder());
+    withStartedStopwatch.process(new Event.Builder().build(), JsonObjectImpl.newBuilder().build(), JsonObjectImpl.newBuilder());
     assertTrue(withStartedStopwatch.duration.toMillis() >= 1);
   }
 
@@ -42,7 +41,7 @@ public class EventProcessorTest {
   void startedStopwatchWithContext() {
     try (ProcessorContext ignored = ProcessorBase.enterContext()) {
       WithStartedStopwatch withStartedStopwatch = new WithStartedStopwatch();
-      withStartedStopwatch.process(EventBuilder.newBuilder().build(), JsonObjectImpl.newBuilder().build(), JsonObjectImpl.newBuilder());
+      withStartedStopwatch.process(new Event.Builder().build(), JsonObjectImpl.newBuilder().build(), JsonObjectImpl.newBuilder());
       assertTrue(withStartedStopwatch.duration.toMillis() >= 1);
     }
   }
@@ -50,7 +49,7 @@ public class EventProcessorTest {
   @Test
   void unstartedStopwatchNoContext() {
     WithUnstartedStopwatch withUnstartedStopwatch = new WithUnstartedStopwatch();
-    withUnstartedStopwatch.process(EventBuilder.newBuilder().build(), JsonObjectImpl.newBuilder().build(), JsonObjectImpl.newBuilder());
+    withUnstartedStopwatch.process(new Event.Builder().build(), JsonObjectImpl.newBuilder().build(), JsonObjectImpl.newBuilder());
     assertTrue(withUnstartedStopwatch.duration.toMillis() >= 20);
   }
 
@@ -58,7 +57,7 @@ public class EventProcessorTest {
   void unstartedStopwatchWithContext() {
     try (ProcessorContext ignored = ProcessorBase.enterContext()) {
       WithUnstartedStopwatch withUnstartedStopwatch = new WithUnstartedStopwatch();
-      withUnstartedStopwatch.process(EventBuilder.newBuilder().build(), JsonObjectImpl.newBuilder().build(), JsonObjectImpl.newBuilder());
+      withUnstartedStopwatch.process(new Event.Builder().build(), JsonObjectImpl.newBuilder().build(), JsonObjectImpl.newBuilder());
       assertTrue(withUnstartedStopwatch.duration.toMillis() >= 20);
     }
   }
